RF: Updated the function locations.
